Create ct-addendum.md
A new Connecticut law requires certain provisions to be included in the contracts between us and the Districts in addition to our existing privacy policy.  This addendum is written as a standalone provision, with the intent that if asked by a district in Connecticut, a rep can point to this provision without having to explain how the language is integrated into the privacy policy.