Fix Private Policy - titles and internal linking
The prior version mixed html into the markdown to add anchors for
linking.  Markdown supports this itself so is not required.

An equivalent fix has been opened as a PR against the monolith but we
need to update here to pass the validation that our policy wording has
not been changed from a legal perspective.